Add files via upload
This Project is done as part of building my personal portfolio using dummy data which is loaded from MySQL. This gives different views of an Organization like Finance, Sales, Marketing and supply chain. This is still in progress and there are many modifications to be done on this. I will upload the final published file here for good User experience in the coming month.